requirements.txt
requests==2.32.3
twscrape==0.17.0
langgraph==0.4.1
anthropic==0.52.0
python-telegram-bot==21.11
playwright==1.51.0
apscheduler==3.10.4
python-dotenv==1.0.1
